Andrew's comments
